pyproject.toml
[project]
name = "learn"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
    # Data analysis core
    "numpy",
    "pandas",
    "matplotlib",
    "scipy",
    "scikit-learn",
    "seaborn",
    # Jupyter for notebooks
    "jupyter",
    "ipython",
    "ipykernel",
    # Markitdown for document conversion
    "markitdown[all]",
    # Utilities
    "python-dotenv",
    "requests",
    "pyyaml>=6.0.3",
    "bibtexparser>=1.4.3",
]
